feat: streamlined firebase inclusion, consolidated into lib/firebase.js. Removed redundant firebase imports from ViewList.js and AddItems.js. Clarified error message from AddItems.js handleSubmit when writing to database
